backend/requirements.txt
flask
python-dotenv
langchain
langchain-ollama
langchain-openai
marshmallow
flask-cors
pinecone-client
langchain-pinecone
langgraph